+++ app/src/main/AndroidManifest.xml
@@ -65,7 +65,7 @@
         </activity>
 
         <service
-            android:name=".MyFirebaseMessagingService">
+            android:name=".data.services.MyFirebaseMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT"/>
             </intent-filter>
